Blocks: Fix BlockEdit hooks not working properly with context
