+++ blocks/block-edit/index.js
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-import classnames from 'classnames';
 import { noop, get } from 'lodash';
 
 /**
@@ -9,16 +8,12 @@ import { noop, get } from 'lodash';
  */
 import { withSelect } from '@wordpress/data';
 import { Component, compose } from '@wordpress/element';
-import { withContext, withFilters, withAPIData } from '@wordpress/components';
+import { withContext, withAPIData } from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
-import {
-	getBlockType,
-	getBlockDefaultClassName,
-	hasBlockSupport,
-} from '../api';
+import Edit from './edit';
 import { BlockEditContextProvider } from './context';
 
 export class BlockEdit extends Component {
@@ -56,34 +51,11 @@ export class BlockEdit extends Component {
 	}
 
 	render() {
-		const { name, attributes = {}, isSelected } = this.props;
-		const blockType = getBlockType( name );
-
-		if ( ! blockType ) {
-			return null;
-		}
-
-		// Generate a class name for the block's editable form
-		const generatedClassName = hasBlockSupport( blockType, 'className', true ) ?
-			getBlockDefaultClassName( name ) :
-			null;
-		const className = classnames( generatedClassName, attributes.className );
-
-		// `edit` and `save` are functions or components describing the markup
-		// with which a block is displayed. If `blockType` is valid, assign
-		// them preferentially as the render value for the block.
-		const Edit = blockType.edit || blockType.save;
-
 		// For backwards compatibility concerns adds a focus and setFocus prop
 		// These should be removed after some time (maybe when merging to Core)
 		return (
 			<BlockEditContextProvider value={ this.state.context }>
-				<Edit
-					{ ...this.props }
-					className={ className }
-					focus={ isSelected ? {} : false }
-					setFocus={ noop }
-				/>
+				<Edit { ...this.props } />
 			</BlockEditContextProvider>
 		);
 	}
@@ -95,7 +67,6 @@ BlockEdit.childContextTypes = {
 };
 
 export default compose( [
-	withFilters( 'blocks.BlockEdit' ),
 	withSelect( ( select ) => ( {
 		postType: select( 'core/editor' ).getEditedPostAttribute( 'type' ),
 	} ) ),
